requirements.txt
djangorestframework
appnope==0.1.0
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-floppyforms==1.6.1
django-toolbelt==0.0.1
factory-boy==2.8.1
Faker==0.7.18
gunicorn==19.7.1
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
traitlets==4.3.2
wcwidth==0.1.7
